resolved tests for resultCache

+++ src/resultCache/file_daemon.py
@@ -1,9 +1,10 @@
 from threading import Thread, Event
 from sched import scheduler
-from typing import Optional, Union
+from typing import Optional
 import time
 import json
 import os
+import shutil
 
 
 class PurgeDaemon(Thread):
@@ -15,15 +16,27 @@ class PurgeDaemon(Thread):
         self._min_expiry_time = min_expiry_time
         self.file_lock = set()
         self.load_schedule()
+        self.create_data_folder()
+
+    def create_data_folder(self):
+        last_slash = self._schedule_location.rfind("/")
+        if last_slash == -1:
+            return
+
+        path = self._schedule_location[:last_slash]
+        if not os.path.exists(path):
+            os.makedirs(path)
 
     def save_schedule(self):
+        # TODO: create dir if not created already
         to_save = []
         for event in self._scheduler.queue:
             delete_time = (event.argument[1], event.time)
-            to_save.append(delete_time)
+            to_save.append(delete_time)        
 
         with open(self._schedule_location, "w") as file:
             json.dump(to_save, file)
+
 
     def load_schedule(self):
         if not os.path.exists(self._schedule_location):
@@ -140,12 +153,7 @@ class FileHandler:
         return min(5 * generation_time + self._min_expiry_time, self._max_expiry_time)
 
     def _put_file(self, username: str, claim_check: str, generation_time: int, data: Optional[str], status: float, errors: Optional[str] = None):
-        if status != 0:
-            eta = generation_time / status
-        else:
-            eta = self._max_expiry_time
-        
-        expiry_time = self.get_expiry_time(2 * eta + generation_time) + time.time()
+        expiry_time = self.get_expiry_time(generation_time) + time.time()
         location = self.get_file_location(username, claim_check, create_folders=True)
         self._purge_daemon.remove_entry(location)
 
@@ -169,7 +177,7 @@ class FileHandler:
     def update_error(self, username: str, claim_check: str, generation_time: int, errors: str):
         self._put_file(username, claim_check, generation_time, None, 1.0, errors)
 
-    def _get_file(self, username: str, claim_check: str) -> dict:
+    def _get_file(self, username: str, claim_check: str, remove: bool = True) -> dict:
         location = self.get_file_location(username, claim_check)
 
         while location in self._purge_daemon.file_lock:
@@ -183,19 +191,21 @@ class FileHandler:
         with open(location, "r") as file:
             data = file.read()
 
-        self._purge_daemon.delete(location, True)
-        self._purge_daemon.remove_entry(location)
+        if remove:
+            self._purge_daemon.delete(location, True)
+            self._purge_daemon.remove_entry(location)
+
         self._purge_daemon.file_lock.remove(location)
 
         return json.loads(data)
 
     def get_status(self, username: str, claim_check: str) -> Optional[float]:
-        file_contents = self._get_file(username, claim_check)
+        file_contents = self._get_file(username, claim_check, False)
 
         if file_contents == None:
             return None
 
-        return file_contents["status"]
+        return str(file_contents["status"])
 
 
     def get(self, username: str, claim_check: str) -> Optional[str]:
@@ -216,5 +226,8 @@ class FileHandler:
         return json.dumps(result)
 
     def delete_all(self, delete_schedule=False):
-        self._purge_daemon.delete_all(delete_schedule)
+        if delete_schedule:
+            shutil.rmtree(self._directory, ignore_errors=True)
+        else:
+            self._purge_daemon.delete_all(False)
